re-arranging deltabot and make it work with latest core-rust high_level_bindings2 branch
