+++ src/deltabot/builtin/echo.py
@@ -14,7 +14,7 @@ def deltabot_configure(bot):
     bot.register_command(
         name="/echo",
         description=_('Echoes back the given text.'),
-        long_description= _(
+        long_description=_(
             'To use it you can simply send a message starting with '
             'the command /echo. For example:\n/echo hello world'),
         func=process_command_echo
@@ -29,4 +29,3 @@ def process_command_echo(command):
         name = message.get_sender_contact().display_name
         text = 'From: {}\nDisplay Name: {}'.format(f, name)
     command.message.chat.send_text(text)
-
